LayoutTests/fast/loader/stateobjects/pushstate-object-types.html

<html>
<head>
<script>

if (window.layoutTestController) {
    layoutTestController.clearBackForwardList();
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

function log(txt)
{
    document.getElementById("logger").appendChild(document.createTextNode(txt + "\n"));
}

function runTest()
{
    beganTest = true;
    history.replaceState("FirstEntry", "Initial entry");
    history.pushState(undefined, "undefined entry");
    history.pushState(null, "null entry");
    history.pushState(false, "false entry");
    history.pushState(true, "true entry");
    history.pushState(42, "Number entry");
    history.pushState("String", "String entry");
    history.pushState(new Date(0), "Date entry");
    history.pushState(new RegExp, "RegExp entry");
    history.pushState(new Array, "Array entry");
    history.pushState(new Object, "Object entry");
    history.pushState(new Error, "Error entry");

/* The following are not yet well enough supported in WebKit to meaningfully test here:
history.pushState([ImageData], "ImageData entry");
history.pushState([Blob], "Blob entry");
history.pushState([File], "File entry");
history.pushState([FileList], "FileList entry");
*/

    history.pushState("BufferEntry", "Last entry");
    history.back();
}

var beganTest = false;

onpopstate = function(event)
{
    // The first time popstate fires, it's because the page has finished loading.
    // Only then can we begin the test.
    if (!beganTest) {
        beganTest = true;
        runTest();
        return;
    }
    
    if (event.state instanceof Date)
        log("State popped - " + event.state.getTime()  + " (type " + typeof event.state + ")");
    else
        log("State popped - " + event.state + " (type " + typeof event.state + ")");

    
    if (event.state != "FirstEntry")
        history.back();
    else if (window.layoutTestController)
        layoutTestController.notifyDone();
}

</script>
<body>
<p>
This test calls pushState with state objects of all the different object types supported by the HTML5 "internal structured cloning algorithm" and makes sure the events contain the expected objects when the states are popped.
</p>
<pre id="logger"></pre>
</body>
</html>
