LayoutTests/fast/loader/stateobjects/pushstate-object-types-expected.txt

This test calls pushState with state objects of all the different object types supported by the HTML5 "internal structured cloning algorithm" and makes sure the events contain the expected objects when the states are popped.

State popped - [object Object] (type object)
State popped - [object Object] (type object)
State popped -  (type object)
State popped - null (type object)
State popped - Wed Dec 31 1969 16:00:00 GMT-0800 (PST) (type object)
State popped - String (type string)
State popped - 42 (type number)
State popped - true (type boolean)
State popped - false (type boolean)
State popped - null (type object)
State popped - undefined (type undefined)
State popped - FirstEntry (type string)
